fix: Parse initiatorname.iscsi instead of re-rendering file. (#18)
* fix: Parse initiatorname.iscsi instead of re-rendering file
If the initiator-dictionary config option is not provided (or hostname isn't present there), 
then a random iqn is generated through /sbin/iscsi-iname to be used for the initiator name.

But with a config-changed event, this iqn is re-generated even if it has already been 
generated and set once before in /etc/iscsi/initiatorname.iscsi. This causes connectivity 
issues as mentioned in the linked bug.

This change checks whether the iqn is already present in the initiatorname.iscsi file and if so, 
uses the same instead of re-generating a new one and re-rendering the file 
(if no configuration is provided via initiator-dictionary). 
It also prevents re-rendering the file if the initiator name exists and it's the same as the one 
configured in the initiator-dictionary.

Logic works as follows:

hostname/ initiator name present in initiator-dictionary
    If initiator name also present in initiatorname.iscsi file:
        If both names are same, don't render file.
        If both names are different, render file with value from dictionary
    If initiatorname.iscsi file doesn't exist, render file with name from initiator-dictionary.

initiator-dictionary absent (or hostname not present there)
    If initiator name present in initiatorname.iscsi file, don't render file.
    If initiatorname.iscsi file doesn't exist, generate random iqn render file with it.

Resolves #3 (lp#2045325)

* tests: add parameterized tests for the above changes.

* refactor: separate method to render the actual initiatorname.iscsi file.

* fix: Patch check_call in test_on_config_changed_fc
Previously, the `_reload_multipathd_service` method was not mocked
as part of this test and executed the `systemctl reload` command
directly on the host running the test.