Add type casting to fix failing mypy lint
The mypy lints were failing since the release of ops 2.17.0.
The problem does not appear to be with ops,
but rather with how type narrowing works with mypy,
and how the charm assigns the unit status in different methods.
See https://github.com/canonical/operator/issues/1401
for more context.

We can fix the issue with no changes to the charm logic
by adding type casting to undo the type narrowing in this case.